backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
google-auth==2.29.0
google-cloud-firestore==2.16.0
google-cloud-translate==3.15.3
google-cloud-aiplatform==1.52.0
google-genai>=0.8.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
requests==2.31.0
firebase-admin==6.5.0
python-multipart==0.0.9
